+++ R/utils-upload.R
@@ -10,15 +10,15 @@
 #' document is completely self-contained, i.e. it does not need external image
 #' files any more, and it is ready to be published online.
 #' @param file the path to the image file to be uploaded
-#' @param key the API key for Imgur (by default uses a key created by Yihui Xie,
-#'   which allows 50 uploads per hour per IP address)
+#' @param key the client id for Imgur (by default uses a client id registered by
+#'   Yihui Xie)
 #' @return A character string of the link to the image; this string carries an
 #'   attribute named \code{XML} which is a list converted from the response XML
 #'   file; see Imgur API in the references.
 #' @author Yihui Xie, adapted from the \pkg{imguR} package by Aaron Statham
-#' @note Please register your own Imgur application to get your API key; you can
-#'   certainly use mine, but this key is in the public domain so everyone has
-#'   access to all images associated to it.
+#' @note Please register your own Imgur application to get your client id; you
+#'   can certainly use mine, but this id is in the public domain so everyone
+#'   has access to all images associated to it.
 #' @references Imgur API: \url{http://api.imgur.com/}; a demo:
 #'   \url{http://yihui.name/knitr/demo/upload/}
 #' @export
@@ -28,17 +28,18 @@
 #' res = imgur_upload(f)
 #' res  # link to original URL of the image
 #' attr(res, 'XML')  # all information
-#' if (interactive()) browseURL(res$links$imgur_page) # imgur page
+#' if (interactive()) browseURL(res)
 #'
 #' ## to use your own key
 #' opts_knit$set(upload.fun = function(file) imgur_upload(file, key = 'your imgur key'))
 #' }
-imgur_upload = function(file, key = '60e9e47cff8483c6dc289a1cd674b40f') {
-  if (is.null(key) || nchar(key) != 32L)
-    stop('The Imgur API Key must be a character string of length 32!')
-  if (!file.exists(file)) stop(file, 'does not exist!')
-  params = list(key = key, image = RCurl::fileUpload(file))
-  res = XML::xmlToList(RCurl::postForm("http://api.imgur.com/2/upload.xml", .params = params))
-  if (is.null(res$links$original)) stop('failed to upload ', file)
-  structure(res$links$original, XML = res)
+imgur_upload = function(file, key = '9f3460e67f308f6') {
+  if (!is.character(key)) stop('The Imgur API Key must be a character string!')
+  res = RCurl::postForm(
+    'https://api.imgur.com/3/image.xml', image = RCurl::fileUpload(file),
+    .opts = RCurl::curlOptions(httpheader = c(Authorization = paste('Client-ID', key)))
+  )
+  res = XML::xmlToList(res)
+  if (is.null(res$link)) stop('failed to upload ', file)
+  structure(res$link, XML = res)
 }