+++ NEWS.md
@@ -34,6 +34,8 @@
 
 ## MINOR CHANGES
 
+- the function `imgur_upload()` uses Imgur API version 3 now; if you are using the key obtained from version 2, you need to register for your own client id: http://api.imgur.com (#439)
+
 - slight improvement of encoding support in `knit()`
 
 # CHANGES IN knitr VERSION 1.0